Fixing unused method linter error
